backend/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
edge-tts
pydantic
opencc-python-reimplemented
